Step5 结算引擎：手动提前结算 + 到期自动兜底，工资/服务费/退款资金闭环
- 金额精度：tb_wallet/tb_wallet_log/tb_job.frozen_amount 改 decimal(10,2)，
  钱包记账改 BigDecimal，半天工资与 10% 服务费精确到分
- 结算主子表：tb_job_settlement(job_id 唯一=幂等锚点) + tb_job_settlement_item
  (逐工人快照)；计薪口径 on+off 都过=1 天/仅 on 过=0.5 天/否则 0，单人封顶
  任务天数×2 保证退款不为负
- 结算核心 settleInternal(事务)：幂等检查→计薪→服务费余额预检→落单→发工资
  →扣 10% 服务费(雇主可用余额另扣)→退冻结余款→报名 1→2→岗位下架+清零冻结
  +删缓存/GEO；@Lazy 自代理保证定时兜底逐岗走事务
- 自动兜底：@EnableScheduling + @Scheduled 扫描 end_time<=now 未结算岗位，
  失败留待下轮重试；手动/自动共用同一核心，重复结算幂等
- 业务封口：编辑/下架、考勤打卡/核销/补记 均查结算单拒绝；互评资格收紧为
  报名 status=2(已完成)；报名已由岗位下架拦截
- 前端：考勤核销页加结算概览(冻结/应发/服务费/退款)+结算按钮，结算后隐藏
  核销按钮；format.js 加 formatMoney
- 修复：结算预览接口未结算时 gross/serviceFee/refund 误返 0，改 toDetailPreview
  回填实时重算金额
- curl E2E：预览/手动结算/幂等/越权/半分精度/冻结清零/自动兜底/0出勤全退/
  服务费不足零副作用/已结算考勤锁定，全部通过

+++ backend/src/main/java/com/linggong/dto/WalletLogDTO.java
@@ -3,6 +3,8 @@ package com.linggong.dto;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 /**
  * 钱包流水返回 DTO（前端「我的钱包」流水列表用）。
  *
@@ -19,11 +21,11 @@ public class WalletLogDTO {
     /** 流水类型：充值/冻结/解冻/工资/服务费 */
     private String type;
 
-    /** 变动金额（正=入账，负=出账） */
-    private Integer amount;
+    /** 变动金额（元，正=入账，负=出账，精确到分） */
+    private BigDecimal amount;
 
-    /** 变动后余额 */
-    private Integer balanceAfter;
+    /** 变动后余额（元，精确到分） */
+    private BigDecimal balanceAfter;
 
     /** 关联业务 id（可为空） */
     private Long bizId;
