+++ backend/src/main/java/com/linggong/entity/Wallet.java
@@ -6,13 +6,14 @@ import com.baomidou.mybatisplus.annotation.TableName;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 /**
  * 虚拟钱包实体，对应表 tb_wallet，与 tb_user 一对一（user_id 唯一）。
  *
- * <p>balance 为可用余额（元，整数）。担保冻结、工资结算都走这个账户：
- * 发岗时从余额里扣冻结金额，结算时从「平台暂扣」里给打工人发工资、给雇主退剩余。
+ * <p>balance 为可用余额（元，精确到分）。担保冻结、工资结算都走这个账户：
+ * 发岗时从余额里扣冻结金额，结算时从平台暂扣资金给打工人发工资、给雇主退剩余。
  * 流水统一记录在 tb_wallet_log。
  */
 @Data
@@ -27,8 +28,8 @@ public class Wallet {
     /** 关联用户 id（雇主/打工人共用） */
     private Long userId;
 
-    /** 可用余额（元，整数） */
-    private Integer balance;
+    /** 可用余额（元，精确到分） */
+    private BigDecimal balance;
 
     /** 创建时间 */
     private LocalDateTime createTime;
